Fixed condition for delete

+++ includes/class-publitio-offloading.php
@@ -34,8 +34,9 @@ class PWPO_Offload
             $this->sizes = $this->publitioApi->_get_all_image_sizes();
             add_action('add_attachment', array($this, 'pwpo_upload_file_to_publitio'));
             if (get_option('publitio_offloading_delete_checkbox') === 'yes') {
-                add_action('deleted_post_meta', array($this, 'pwpo_delete_file_from_publitio'), 10, 4);
+                add_action('delete_post_meta', array($this, 'pwpo_delete_file_from_publitio'), 10, 4);
             }
+
             add_filter('the_content', array($this, 'pwpo_update_offloading_images_src'), 10);
             add_filter('wp_calculate_image_srcset', array($this, 'pwpo_calculate_image_offloading_srcset'), 10, 5);
             add_filter('image_downsize', array($this, 'pwpo_filter_image_downsize'), 10, 3);
@@ -196,11 +197,13 @@ class PWPO_Offload
      */
     public function pwpo_delete_file_from_publitio($deleted_meta_ids, $post_id, $meta_key, $only_delete_these_meta_values)
     {
-        $publitioMeta = get_post_meta($post_id, 'publitioMeta', true);
-        if ($publitioMeta) {
-            $responseShow = $this->publitioApi->showFile( $publitioMeta['id']);
-            if ($responseShow->success === true) {
-                $this->publitioApi->deleteFileFromPublitio($publitioMeta['id']);
+        if($meta_key === 'publitioMeta') {
+            $publitioMeta = $only_delete_these_meta_values;
+            if ($publitioMeta) {
+                $responseShow = $this->publitioApi->showFile( $publitioMeta['id']);
+                if ($responseShow->success === true) {
+                    $this->publitioApi->deleteFileFromPublitio($publitioMeta['id']);
+                }
             }
         }
     }
